app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Илья Ватютов",email = "vatut007@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "sqlmodel (>=0.0.32,<0.0.33)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "orjson (>=3.11.7,<4.0.0)",
]

[tool.poetry]
packages = [{include = "app", from = "src"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
